Drop support for pipeline version 0
Lets us get rid of `getCompletion` and simplify a fair amount of code surrounding it.
